+++ admin/src/index.ts
@@ -42,27 +42,29 @@ async function main() {
     const test_id = document.getElementById("new_test_id")! as HTMLInputElement;
     const test_name = document.getElementById("new_test_name")! as HTMLInputElement;
     const test_article = document.getElementById("new_test_test_article")! as HTMLSelectElement;
-    const gse_article = document.getElementById("new_test_gse_article")! as HTMLInputElement;
+    const gse_article = document.getElementById("new_test_gse_article")! as HTMLSelectElement;
+    const trim_to = document.getElementById("new_test_trim")! as HTMLInputElement;
     const entries = document.getElementsByClassName("url_list_div")!.length;
+    let trim_to_s: number = 0;
+    let isnum = /^\d+$/.test(trim_to.value);
+    if (isnum) {
+      trim_to_s = parseInt(trim_to.value);
+    }
     let url_pairs: string[] = [];
     for (let i = 0; i < entries; i++) {
       const element = document.getElementById("url_list_div_textinput_" + i)! as HTMLInputElement;
       url_pairs.push(element.value);
     }
-    console.log({
+    const payload = {
       test_name: test_name.value,
       test_id: test_id.value,
       test_article: test_article.value,
       gse_article: gse_article.value,
       url_pairs: url_pairs,
-    });
-    createTest({
-      test_name: test_name.value,
-      test_id: test_id.value,
-      test_article: test_article.value,
-      gse_article: gse_article.value,
-      url_pairs: url_pairs,
-    }).then(async (result) => {
+      trim_to_s: trim_to_s.toString(),
+    };
+    console.log(payload);
+    createTest(payload).then(async (result) => {
       const data: any = result.data;
       // const message = data.csv_fields;
       console.log(data);
@@ -70,7 +72,7 @@ async function main() {
       // csvButton.innerHTML = check_mark;
       // await delay(1500);
       // csvButton.innerHTML = '<span class="material-symbols-outlined">table_view</span>';
-    updateStatus(loadingStatus.DONE);
+      updateStatus(loadingStatus.DONE);
     });
   });
   generateArticlePanel(articleType.TEST);